mysite/requirements.txt
attrs==17.4.0
arrow==0.12.1
awsebcli==3.10.6
blessed==1.14.2
boto==2.48.0
boto3==1.4.7
botocore==1.7.2
bz2file==0.98
celery==4.0.2	
cement==2.8.2	
click==6.7		
colorama==0.3.9
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.7.1
Django==1.11.13
django-unixdatetimefield==0.1.6
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==0.12.2
funcsigs==1.0.2
funcy==1.10.1
future==0.16.0
futures==3.1.1
fuzzywuzzy==0.15.1
gensim==3.3.0
google-cloud-language==1.0.2
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
joblib==0.11
MarkupSafe==1.0
matplotlib==1.5.1
mpld3==0.3
murmurhash==0.26.4
mysql==0.0.1
mysql-connector-python==8.0.6
MySQL-python==1.2.5
nltk==3.2.1
numexpr==2.6.4
numpy==1.14.0
oauthlib==2.0.6
pandas==0.18.1
pathlib==1.0.1
pathspec==0.5.0
plac==0.9.6
pluggy==0.6.0
preshed==1.0.0
py==1.5.2
pyLDAvis==2.1.1
PyMySQL==0.8.0
pyreadline==2.1
pytest==3.4.0
python-dateutil==2.6.1
pytz==2016.6.1
PyYAML==3.12
requests==2.11.1
requests-oauthlib==0.8.0
s3transfer==0.1.11
scikit-learn==0.19.1
scipy==1.0.0
scrapy==1.3.3
scrapy-djangoitem==1.1.1
scrapyd==1.2.0
selenium==2.53.6
semantic-version==2.5.0
shub==2.8.0
six==1.10.0
smart-open==1.5.6
SpeechRecognition==3.7.1
tabulate==0.7.5
termcolor==1.1.0
thinc==6.5.2
toolz==0.8.2
tqdm==4.15.0
tweepy==3.5.0
tweet-preprocessor==0.5.0
urllib3==1.17
wcwidth==0.1.7
websocket-client==0.44.0
Werkzeug==0.12.2
wrapt==1.10.11